Rewrite fork to support page sharing across forks (L5-Ex8)
Completely rewrite fork's duppage(). Fork's code is much better now, it was luck that it worked before.
The change is mostly in the way it copies the pages. Also added support for pages that have PTE_SHARE enabled.

"We have defined a new PTE_SHARE bit in inc/lib.h. This bit is one of the three PTE bits that are marked "available for software use" in the Intel and AMD manuals. We will establish the convention that if a page table entry has this bit set, the PTE should be copied directly from parent to child in both fork and spawn. Note that this is different from marking it copy-on-write: as described in the first paragraph, we want to make sure to share updates to the page."