Bugfix: corrects fork behavior with read-only pages
Fork didn't support read-only (without PTE_W) on user space, because it was led to believe that every page in user space shoulded have PTE_W enabled.
This happened because load_icode put all the pages with PTE_U | PTE_W, but spawn actually set the pages permissions according to the specified in the ELF.
With this, the code did panic when forking processes that were loaded via spawn. This bugfix solves this issue by simply sharing the read-only pages.